requirements.txt
flask>=2.0
flask-cors>=3.0
openai>=1.0
anthropic>=0.18
requests>=2.28
python-dotenv>=1.0
pydub>=0.25
